[feature] use sessions to identify user making api calls
