+++ server/models/index.js
@@ -99,6 +99,41 @@ module.exports = {
     }
   },
 
+  review: {
+    get: function(locationId, callback){
+      db.Review.findAll({
+        where: {
+          location_id: locationId
+        }
+      }).then(function(reviews){
+        var queries = [];
+        _.each(reviews, function(review){
+          queries.push(db.User.findById(review.user_id).then(function(user){
+            review.dataValues.author = user.name;
+          }));
+        })
+        Promise.all(queries)
+        .then(function(){
+          callback(reviews);
+        })
+      })
+    },
+    post: function(review, callback){
+      console.log("+++line 122 what is the review", review)
+      db.Review.create({
+        location_id: review.location_id,
+        user_id: review.user_id,
+        body: review.body,
+        rating: review.rating
+      }).then(function(postedReview){
+        callback(postedReview);
+      }).catch(function(error){
+        console.log("+++line 131 models posting review failed")
+        console.log(error)
+      })
+    }
+  },
+
   login: {
     post: function (username, password, callback) {
       db.User.find({
